chore: extract credential cache to shared utility in ociauthconfig
Closes #49

Extract the singleflight + TTL credential-helper caching logic from
internal/backend/remote-state/oras/backend.go into a new shared type
ociauthconfig.CachedCredentialsLookupEnv in
internal/command/cliconfig/ociauthconfig/cached_credentials_env.go.

Apply the cache to cmd/ghoten/oci_distribution.go via a process-scoped
ociCredLookupEnv so that docker-style credential helpers are invoked at
most once per registry per cache window (default 5 min) during module
and provider installation.

Changes:
- ociauthconfig: new CachedCredentialsLookupEnv with TTL + singleflight;
  exported DefaultCredentialHelperCacheTTL constant
- ociauthconfig: move credential cache unit tests here (6 tests), add
  TestCachedCredentialsLookupEnv_ZeroTTLDisablesCache
- oras/backend.go: replace inline cache impl with shared utility; remove
  ~120 lines of duplicate code; remove unused sync import
- oras/backend_test.go: remove credential cache tests (now in ociauthconfig)
- cmd/ghoten/oci_distribution.go: add process-scoped ociCredLookupEnv
