+++ internal/backend/remote-state/oras/backend_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"crypto/x509"
 	"encoding/pem"
-	"fmt"
 	"io"
 	"net/http"
 	"net/http/httptest"
@@ -16,7 +15,6 @@ import (
 	"time"
 
 	"github.com/vmvarela/ghoten/internal/backend"
-	"github.com/vmvarela/ghoten/internal/command/cliconfig/ociauthconfig"
 	"github.com/vmvarela/ghoten/internal/configs"
 	"github.com/vmvarela/ghoten/internal/encryption"
 	"github.com/zclconf/go-cty/cty"
@@ -251,242 +249,6 @@ func (rt *headerCapturingRoundTripper) RoundTrip(req *http.Request) (*http.Respo
 		Header:     make(http.Header),
 		Request:    req,
 	}, nil
-}
-
-type countingLookupEnv struct {
-	mu    sync.Mutex
-	calls int
-
-	result ociauthconfig.DockerCredentialHelperGetResult
-	err    error
-}
-
-func (e *countingLookupEnv) QueryDockerCredentialHelper(ctx context.Context, helperName string, serverURL string) (ociauthconfig.DockerCredentialHelperGetResult, error) {
-	e.mu.Lock()
-	e.calls++
-	e.mu.Unlock()
-	return e.result, e.err
-}
-
-func (e *countingLookupEnv) Calls() int {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	return e.calls
-}
-
-func TestCachedDockerCredentialHelperEnv_CachesSuccessWithinTTL(t *testing.T) {
-	inner := &countingLookupEnv{
-		result: ociauthconfig.DockerCredentialHelperGetResult{
-			ServerURL: "https://example.com",
-			Username:  "u",
-			Secret:    "s",
-		},
-	}
-	env := newCachedDockerCredentialHelperEnv(inner, time.Minute)
-	base := time.Unix(100, 0)
-	env.now = func() time.Time { return base }
-
-	ctx := context.Background()
-	_, err := env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	_, err = env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	if got, want := inner.Calls(), 1; got != want {
-		t.Fatalf("inner calls = %d, want %d", got, want)
-	}
-}
-
-func TestCachedDockerCredentialHelperEnv_ExpiresAfterTTL(t *testing.T) {
-	inner := &countingLookupEnv{
-		result: ociauthconfig.DockerCredentialHelperGetResult{
-			ServerURL: "https://example.com",
-			Username:  "u",
-			Secret:    "s",
-		},
-	}
-	env := newCachedDockerCredentialHelperEnv(inner, 10*time.Second)
-	base := time.Unix(100, 0)
-	now := base
-	env.now = func() time.Time { return now }
-
-	ctx := context.Background()
-	_, err := env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	now = base.Add(11 * time.Second)
-	_, err = env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	if got, want := inner.Calls(), 2; got != want {
-		t.Fatalf("inner calls = %d, want %d", got, want)
-	}
-}
-
-func TestCachedDockerCredentialHelperEnv_CachesNotFoundErrorWithinTTL(t *testing.T) {
-	notFoundErr := ociauthconfig.NewCredentialsNotFoundError(context.Canceled) // any wrapped error is fine
-	inner := &countingLookupEnv{err: notFoundErr}
-	env := newCachedDockerCredentialHelperEnv(inner, time.Minute)
-	base := time.Unix(100, 0)
-	env.now = func() time.Time { return base }
-
-	ctx := context.Background()
-	_, err := env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-	if err == nil || !ociauthconfig.IsCredentialsNotFoundError(err) {
-		t.Fatalf("expected not-found error, got %v", err)
-	}
-	_, err = env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-	if err == nil || !ociauthconfig.IsCredentialsNotFoundError(err) {
-		t.Fatalf("expected not-found error, got %v", err)
-	}
-
-	if got, want := inner.Calls(), 1; got != want {
-		t.Fatalf("inner calls = %d, want %d", got, want)
-	}
-}
-
-func TestCachedDockerCredentialHelperEnv_UnexpectedErrorUsesShortTTL(t *testing.T) {
-	// Unexpected errors (not "credentials not found") should be cached with
-	// a short TTL so that transient failures recover quickly.
-	transientErr := fmt.Errorf("connection refused")
-	inner := &countingLookupEnv{err: transientErr}
-	env := newCachedDockerCredentialHelperEnv(inner, time.Minute)
-	base := time.Unix(100, 0)
-	now := base
-	env.now = func() time.Time { return now }
-
-	ctx := context.Background()
-	_, err := env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-	if err == nil {
-		t.Fatalf("expected error, got nil")
-	}
-	// Same time: should be cached.
-	_, _ = env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-	if got, want := inner.Calls(), 1; got != want {
-		t.Fatalf("expected 1 inner call (cached), got %d", got)
-	}
-
-	// Advance past the short error TTL (errorCacheTTL = 10s) but not the full TTL (60s).
-	now = base.Add(11 * time.Second)
-	_, err = env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-	if err == nil {
-		t.Fatalf("expected error, got nil")
-	}
-	if got, want := inner.Calls(), 2; got != want {
-		t.Fatalf("expected 2 inner calls after short TTL expired, got %d", got)
-	}
-}
-
-func TestCachedDockerCredentialHelperEnv_SingleflightDeduplication(t *testing.T) {
-	// Verify that concurrent calls for the same key are deduplicated.
-	gate := make(chan struct{})
-	inner := &countingLookupEnv{
-		result: ociauthconfig.DockerCredentialHelperGetResult{
-			ServerURL: "https://example.com",
-			Username:  "u",
-			Secret:    "s",
-		},
-	}
-	// Wrap inner to block until gate is released.
-	blocking := &blockingLookupEnv{inner: inner, gate: gate}
-	env := newCachedDockerCredentialHelperEnv(blocking, time.Minute)
-	env.now = func() time.Time { return time.Unix(100, 0) }
-
-	ctx := context.Background()
-	const concurrency = 5
-	results := make(chan error, concurrency)
-
-	for i := 0; i < concurrency; i++ {
-		go func() {
-			_, err := env.QueryDockerCredentialHelper(ctx, "example", "https://example.com")
-			results <- err
-		}()
-	}
-
-	// Let all goroutines reach the singleflight barrier, then release.
-	time.Sleep(50 * time.Millisecond)
-	close(gate)
-
-	for i := 0; i < concurrency; i++ {
-		if err := <-results; err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-	}
-
-	// Despite 5 concurrent callers, only 1 actual helper invocation should have occurred.
-	if got := inner.Calls(); got != 1 {
-		t.Fatalf("expected 1 inner call (singleflight), got %d", got)
-	}
-}
-
-// blockingLookupEnv wraps a countingLookupEnv and blocks until gate is closed.
-type blockingLookupEnv struct {
-	inner *countingLookupEnv
-	gate  <-chan struct{}
-}
-
-func (e *blockingLookupEnv) QueryDockerCredentialHelper(ctx context.Context, helperName string, serverURL string) (ociauthconfig.DockerCredentialHelperGetResult, error) {
-	<-e.gate
-	return e.inner.QueryDockerCredentialHelper(ctx, helperName, serverURL)
-}
-
-// slowLookupEnv blocks until its context is cancelled, simulating a hung
-// credential helper process.
-type slowLookupEnv struct{}
-
-func (e *slowLookupEnv) QueryDockerCredentialHelper(ctx context.Context, helperName string, serverURL string) (ociauthconfig.DockerCredentialHelperGetResult, error) {
-	<-ctx.Done()
-	return ociauthconfig.DockerCredentialHelperGetResult{}, ctx.Err()
-}
-
-func TestCachedDockerCredentialHelperEnv_TimeoutErrorUsesShortTTL(t *testing.T) {
-	// Use a very short TTL so that a simulated timeout expires quickly.
-	const shortCacheTTL = time.Minute
-
-	now := time.Now()
-	clock := now
-
-	env := &cachedDockerCredentialHelperEnv{
-		inner: &slowLookupEnv{},
-		ttl:   shortCacheTTL,
-		now:   func() time.Time { return clock },
-	}
-	env.cache = make(map[dockerCredentialHelperCacheKey]dockerCredentialHelperCacheEntry)
-	env.inflight = make(map[dockerCredentialHelperCacheKey]*singleflightCall)
-
-	// Use a context with a very short timeout to simulate the helper hanging.
-	ctx, cancel := context.WithTimeout(context.Background(), 50*time.Millisecond)
-	defer cancel()
-
-	_, err := env.QueryDockerCredentialHelper(ctx, "test-helper", "https://example.com")
-	if err == nil {
-		t.Fatal("expected an error from a timed-out credential helper, got nil")
-	}
-
-	// The error must NOT be a CredentialsNotFoundError — it is a context error.
-	if ociauthconfig.IsCredentialsNotFoundError(err) {
-		t.Fatalf("expected a non-not-found error (timeout), got a CredentialsNotFoundError: %v", err)
-	}
-
-	// A timeout error should be cached with errorCacheTTL (10s), not the full TTL.
-	// Advance the clock by errorCacheTTL + 1s and verify the cache entry has expired.
-	clock = now.Add(errorCacheTTL + time.Second)
-
-	inner2 := &countingLookupEnv{err: fmt.Errorf("second call")}
-	env.inner = inner2
-
-	_, _ = env.QueryDockerCredentialHelper(context.Background(), "test-helper", "https://example.com")
-	if got := inner2.Calls(); got != 1 {
-		t.Fatalf("expected cache to have expired after errorCacheTTL, but inner was called %d times (want 1)", got)
-	}
 }
 
 // writeTempFile writes data to a new file in t.TempDir() and returns the path.
